more options
adds a "more options" option to reduce clutter, while including an advanced options option for those who wish to mess with it. this commit also removes the download folder in favour of the new temp folder. you can also now input 0 as a file size and it will not shrink at all